requirements.txt
django-cors-middleware==1.4.0
Pygments==2.2.0
redgreenunittest==0.1.1
gunicorn==19.9.0
python-http-client==3.2.1
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.3.0
Django==3.0.3
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
importlib-metadata==1.5.0
jmespath==0.9.4
kombu==4.6.7
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==2.2.0
requests==2.23.0
psycopg2==2.8.4
django-filter==2.2.0
sendgrid==6.1.2
sentry-sdk==0.14.3

